requirements.txt
numpy==1.26.4
scipy==1.15.3
sentencepiece==0.1.99
soundfile==0.12.1
tiktoken==0.6.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
train==0.0.5
transformers==4.57.3
wandb==0.16.6
WeTextProcessing==1.0.3
x_transformers==1.42.26
zhconv==1.4.3
zhon==2.1.1

inflect==7.3.1
onnxruntime_gpu==1.19.0
openai_whisper==20231117
contractions==0.1.73
debugpy==1.8.0
deepspeed==0.14.5
einops==0.8.1
emoji==2.14.1
evaluation==0.0.2
fastapi==0.123.9
funasr==1.1.6
gradio==6.0.2
HyperPyYAML==1.2.2
jieba==0.42.1
jiwer==4.0.0
librosa==0.9.2
matplotlib==3.8.2
peft==0.7.0
pronouncing==0.2.0
pydantic==2.12.4
pydub==0.25.1
pypinyin==0.44.0
PyYAML==6.0.3
regex==2023.12.25
Requests==2.32.5
safetensors==0.7.0
soxr